moved db access method to model, maybe I should make a router too
